* combigrid: Fix warnings in examples/combi_classic.cpp.
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@4184 4eea3252-f0fb-4393-894d-40516dce545b

+++ combigrid/examples/combi_classic.cpp
@@ -32,8 +32,8 @@ int main() {
   /*print the combined subspaces and coefficients*/
   std::cout << "combined subspaces: " << std::endl;
 
-  for (int i = 0; i < coeffs.size(); i++) {
-    for (int j = 0; j < levels[i].size(); j++) {
+  for (size_t i = 0; i < coeffs.size(); i++) {
+    for (size_t j = 0; j < levels[i].size(); j++) {
       std::cout << levels[i][j] << '\t';
     }
